requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.4
ultralytics==8.0.186
easyocr==1.7.0
supabase==1.0.3
python-dotenv==1.0.0
pillow==10.0.1
torch==2.0.1
torchvision==0.15.2